chore: Moved types from lib.rs to separate modules
